requirements.txt
boto3>=1.28.0
PyYAML>=6.0
requests>=2.31.0